Move Network model to commons
