+++ universal-login-commons/src/index.ts
@@ -89,3 +89,5 @@ export {isDataForFunctionCall} from './core/utils/isDataForFunctionCall';
 export {multiplyBy150Percent} from './core/utils/multiplyBy150Percent';
 export {isClassName} from './react/isClassName';
 export {raise} from './core/utils/raise';
+export {CurrencyValue} from './core/models/CurrencyValue';
+export {Network, asNetwork} from './core/models/Network';
