+++ universal-login-web3/src/services/setupStrategies.ts
@@ -1,7 +1,7 @@
 import {Web3ProviderFactory} from '../models/Web3ProviderFactory';
-import {getConfigForNetwork, Network} from '../config';
+import {getConfigForNetwork} from '../config';
 import {ULWeb3Provider} from '../ULWeb3Provider';
-import {ApplicationInfo} from '@unilogin/commons';
+import {ApplicationInfo, Network} from '@unilogin/commons';
 import {StorageService, BrowserChecker} from '@unilogin/react';
 import UniLoginLogo from '../ui/assets/U.svg';
 import MetamaskLogo from '../ui/assets/MetaMaskLogoTitle.svg';
